update: 프로필 수정, 회원탈퇴, 모달 토글 기능 구현
