+++ src/pages/MyPage/MyPageContent.jsx
@@ -1,0 +1,5 @@
+const MyPageContent = () => {
+  return <div></div>;
+};
+
+export default MyPageContent;
